requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pyjwt==2.8.0
slowapi==0.1.9
psycopg2-binary==2.9.9